Move order commands files into order package
