+++ src/main/java/seedu/address/logic/commands/order/DeleteCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.order;
 
 import static java.util.Objects.requireNonNull;
 
@@ -7,6 +7,8 @@ import java.util.List;
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.CommandHistory;
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.order.Order;
@@ -14,7 +16,7 @@ import seedu.address.model.order.Order;
 /**
  * Deletes an order identified using it's displayed index from the order book.
  */
-public class DeleteCommand extends Command {
+public class DeleteCommand extends OrderCommand {
 
     public static final String COMMAND_WORD = "delete";
