+++ src/main/java/seedu/address/logic/commands/order/ListCommand.java
@@ -1,15 +1,17 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.order;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_ORDERS;
 
 import seedu.address.logic.CommandHistory;
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
 
 /**
  * Lists all orders in the order book to the user.
  */
-public class ListCommand extends Command {
+public class ListCommand extends OrderCommand {
 
     public static final String COMMAND_WORD = "list";
