+++ src/main/java/seedu/address/logic/parser/order/SelectCommandParser.java
@@ -1,9 +1,11 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.order;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.SelectCommand;
+import seedu.address.logic.commands.order.SelectCommand;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
