+++ src/test/java/seedu/address/testutil/OrderUtil.java
@@ -5,7 +5,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_FOOD;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
 
-import seedu.address.logic.commands.AddCommand;
+import seedu.address.logic.commands.order.AddCommand;
 import seedu.address.model.order.Order;
 
 /**
