+++ src/test/java/seedu/address/ui/CommandBoxTest.java
@@ -11,7 +11,7 @@ import guitests.guihandles.CommandBoxHandle;
 import javafx.scene.input.KeyCode;
 import seedu.address.logic.Logic;
 import seedu.address.logic.LogicManager;
-import seedu.address.logic.commands.ListCommand;
+import seedu.address.logic.commands.order.ListCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
