+++ src/test/java/systemtests/OrderBookSystemTest.java
@@ -33,9 +33,9 @@ import seedu.address.MainApp;
 import seedu.address.TestApp;
 import seedu.address.commons.core.EventsCenter;
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.ClearCommand;
-import seedu.address.logic.commands.ListCommand;
-import seedu.address.logic.commands.SelectCommand;
+import seedu.address.logic.commands.order.ClearCommand;
+import seedu.address.logic.commands.order.ListCommand;
+import seedu.address.logic.commands.order.SelectCommand;
 import seedu.address.model.Model;
 import seedu.address.model.OrderBook;
 import seedu.address.testutil.TypicalOrders;
